feat: Delete 'get_coordinates_and_elevation' function from imgw.py and 'd_imgw_coordinates' method from the WalterLieth class in walter_lieth.py. Add a new argument to 'get_meteorological_data' in imgw.py, which allows to get station coordinates already at this stage
Downloading coordinates for the IMGW stations used to be executed with the additional function from imgw.py (the function, which used BeautifulSoup, so sometimes it was quite lengthy to get coordinates). Now it is possible to get station coordinates in 'get_meteorological_data' function. The coordinate data is retrieved from a local file, so the whole process is much faster.