fix: Sort dataframe by a list, not a list inside another list. Return station, year and month columns when downloading the date for the entire country.
The 'download_meteo_data' function from climex_scraping.py used to sort the order of the columns by a list inside another list, which caused errors every time the feature of the function was used. Sorting contributed to omission of the year and month columns when the dataframe was returned, so now it does not exclude the above columns.