fix: Take 'years_range' argument from the method, not from the attributes of the class (self.years_range)
Data downloading by the 'd_imgw_data' method with a daily interval caused errors. The method used to take a value for the 'years_range' argument from the attributes of the class, where 'self.years_range' is None by default. Now a value is taken from the method arguments, so if user specifies valid years range, the data will be downloaded.