fix: Change the way of testing graphs
The graphs used to be tested by file sizes comparision. However, it was problematic while the platform or package's version (matplotlib.pyplot) was changed. From now on, testing assumes that the matplotlib library will draw graphs correctly and only checks if no errors occured during the drawing process. This approach is the most flexible, the fastest and the less complicated. What is more, the testing process does not require any additional files (images) to compare, so the cloupy's weight is not unnecessarily disturbed.