fix: Change the way of drawing shapes
From now counting the distance between two consecutive points is not necessary. Now the function looks for the starting point, and if it is found, the shape is being drawn. After that, shape drawing process is finished and starts again for the next shape. It is no longer necessary to count the distance between two consecutive points to aboid drawing lines across the whole graph. Now the function recognises seperated shapes in more efficient and precise way