refactor: Change the color for creating mask
A color for creating mask was changed (from now it is much more similar to black - the color of the boundaries), so pixels which lay close to the boundaries do not have unwanted color. The masking color used to be green, so it was significantly different from black, and you could see unwanted pixel colors close to the boundaries.