fix: Extrapolate into zoomed in area by default, allow to extrapolate over the zoomed in area if necessary
When user zooms in the map, the values will be extrapolated to the zoomed in area corners by default. It is handy when user zooms in on a country which has overseas territories - in this case, if the data is only for the zoomed in country, extrapolation to the corners of the whole shapefile could result in incorrect interpolation. However, the user may want to zoom into a specific part of the map without changing the extrapolation corners, so it is still possible by changing the extrapolate_into_zoomed_area argument to False.