refactor: Use lower dpi when not saving the interpolation map
When the interpolation map is not to be saved, lower dpi value will be used for displaying previewed interpolation map. When the interpolation map is to be saved, higher dpi value (300) will be used for saving and displaying the map. Using lower dpi value for previewed maps saves significantly time, so a working process with the class is much more efficient.