fix: Fix passing int to the 'station_name' argument
The station_name argument can be an int (which actually is a station's WMO ID). If the user wants to search by the WMO ID, the 'station_name_is_wmo_id' argument must be set to True. It seems like in the developing process changing the 'station_name' argument to the string was deleted which caused problems with the searching by the WMO ID (the code tried to use string's methods on the int). Changing the 'station_name' argument to the string in the beginning of the function has just been implemented again.