fix: Change the way in which choosing the global style works
Choosing the global style worked for all diagrams and maps, but the global styles were adapted only for the Walter-Lieth diagram. When the user tried to set global style for the interpolation map, the global changes had incorrect impact on the way in which the map was drawn. From now, choosing the global style changes only the most general settings and the information about the chosen global style is stored in a .txt file. Every graph/map checks which global style was chosen and fits the other - more specific - parameters respectively to the chosen global style.