fix: Allow passing strings to the 'keywords' argument
When user wanted to filter columns by the given keywords, the 'keywords' argument had to be a list - if it was a single string, the errors occured. From now, at the beginning of the functions (which use 'keywords' argument) the 'keywords' argument is transformed to be a list if it is a string.