fix: Fill everything under the 0 line level without the negative impact on the wet periods hatch
In the Walter-Lieth diagram, everything under the 0 line level must be white - there can not be any hatch of wet periods or dry periods. To this moment, the condition was met, but it distorted slightly the hatches. From now, the condition does not affect the hatches - firstly the hatches of wet period and dry period are drawn, and after that a white layer covers any hatch which is under the 0 line level.