fix: Cover the place between the 0 line with white color even when the 'freeze_rectangles' argument is set to Falsle
The bottom boundary for the 'fill_between' method was the 'y_value_for_closing_bottom_rectangles' and it worked fine when the freeze rectangles were to be drawn. However, when the bottom rectangles were not to be drawn, the 'y_value_for_closing_bottom_rectangles' was None which resulted in errors. From now, when the 'y_value_for_closing_bottom_rectangles' is None, the method takes the minimum value of the 'temp_axis'.